Fix SaleLadger (collectly implement with current database)

+++ src/com/core/database/SaleLadger.java
@@ -4,36 +4,57 @@ import java.util.ArrayList;
 import java.util.List;
 
 import android.content.Context;
+import android.util.Log;
 
 import com.core.Sale;
-import com.database.CustomerBookDB;
 import com.database.SaleLadgerDB;
 
 public class SaleLadger {
 	private SaleLadgerDB db;
-	private List<Sale> sales;
-	public SaleLadger(){
-		sales = new ArrayList<Sale>();
-	}
 	
-	public void add(Context con,Sale sale){
-		db = new SaleLadgerDB(con);
-		db.insert(sale);
+	public Sale getByID(Context context , int id){
+		db = new SaleLadgerDB(context);
+		Sale s = db.findByID(id);
 		db.close();
-		sales.add(sale);
+		return s;
 	}
 	
+	public Sale add(Context con,Sale sale){
+		db = new SaleLadgerDB(con);
+		Sale s = db.insert(sale);
+		db.close();
+		return s;
+	}
+	
+	/*
+	 * have to have payment id items customer.
+	 */
 	public boolean remove(Context con,Sale sale){
+		if(sale.getPayment() == null || sale.getItems() == null || sale.getCustomer() == null) return false;
 		db = new SaleLadgerDB(con);
 		db.delete(sale);
 		db.close();
-		return sales.remove(sale);
+		return true;
 	}
 	
-	public Sale[] getAllSale(Context con){
+	public List<Sale> getAllSales(Context con){
 		db = new SaleLadgerDB(con);
-		Sale[] x = db.findAll();
+		List<Sale> x = db.findAll();
 		db.close();
 		return x;
+	}
+	
+	public boolean isContains(Context con , Sale sale){
+		db = new SaleLadgerDB(con);
+		Sale s = db.findByID(sale.getID());
+		db.close();
+		return s !=null;
+	}
+	
+	public int getQuantity(Context con){
+		db = new SaleLadgerDB(con);
+		int s = db.findAll().size();
+		db.close();
+		return s;
 	}
 }
